+++ integration-tests/api/__tests__/batch-jobs/product/ff-sales-channel.js
@@ -118,17 +118,16 @@ describe("Product import - Sales Channel", () => {
 
     const productsResponse = await api.get("/admin/products", adminReqConfig)
 
-    expect(productsResponse.data.count).toBe(2)
+    expect(productsResponse.data.count).toBe(1)
     expect(productsResponse.data.products).toEqual([
       expect.objectContaining({
-        id: "O6S1YQ6mKm",
         title: "Test product",
-        description: "test-product-description-1",
+        description:
+          "Hopper Stripes Bedding, available as duvet cover, pillow sham and sheet.\\n100% organic cotton, soft and crisp to the touch. Made in Portugal.",
         handle: "test-product-product-1",
         variants: [
           expect.objectContaining({
             title: "Test variant",
-            product_id: "O6S1YQ6mKm",
             sku: "test-sku-1",
           }),
         ],
@@ -142,25 +141,6 @@ describe("Product import - Sales Channel", () => {
             is_disabled: false,
           }),
         ],
-      }),
-      expect.objectContaining({
-        id: "5VxiEkmnPV",
-        title: "Test product",
-        description: "test-product-description",
-        handle: "test-product-product-2",
-        variants: [
-          expect.objectContaining({
-            title: "Test variant",
-            product_id: "5VxiEkmnPV",
-            sku: "test-sku-2",
-          }),
-          expect.objectContaining({
-            title: "Test variant",
-            product_id: "5VxiEkmnPV",
-            sku: "test-sku-3",
-          }),
-        ],
-        sales_channels: [],
       }),
     ])
   })