+++ integration-tests/api/__tests__/taxes/manual-taxes.js
@@ -25,7 +25,7 @@ describe("Manual Cart Taxes", () => {
   beforeAll(async () => {
     const cwd = path.resolve(path.join(__dirname, "..", ".."))
     dbConnection = await initDb({ cwd })
-    medusaProcess = await setupServer({ cwd })
+    medusaProcess = await setupServer({ cwd, verbose: false })
   })
 
   afterAll(async () => {
@@ -76,7 +76,7 @@ describe("Manual Cart Taxes", () => {
 
     const response = await api.get("/store/carts/test-cart")
     expect(response.status).toEqual(200)
-    expect(response.data.cart.tax_total).toEqual(null)
+    expect(response.data.cart.tax_total).toEqual(0)
     expect(response.data.cart.total).toEqual(100)
   })