+++ packages/medusa/src/api/routes/store/carts/delete-line-item.ts
@@ -1,7 +1,6 @@
 import { EntityManager } from "typeorm"
 import { defaultStoreCartFields, defaultStoreCartRelations } from "."
 import { CartService } from "../../../../services"
-import { decorateLineItemsWithTotals } from "./decorate-line-items-with-totals"
 
 /**
  * @oas [delete] /carts/{id}/line-items/{line_id}
@@ -67,11 +66,9 @@ export default async (req, res) => {
     }
   })
 
-  const cart = await cartService.retrieve(id, {
+  const data = await cartService.retrieveWithTotals(id, {
     select: defaultStoreCartFields,
     relations: defaultStoreCartRelations,
   })
-  const data = await decorateLineItemsWithTotals(cart, req)
-
   res.status(200).json({ cart: data })
 }