+++ packages/medusa/src/api/routes/store/products/list-products.ts
@@ -141,7 +141,16 @@ import { IsType } from "../../../../utils/validators/is-type"
  *             products:
  *               type: array
  *               items:
- *                 $ref: "#/components/schemas/product"
+ *                 allOf:
+ *                   - $ref: "#/components/schemas/product"
+ *                   - type: object
+ *                     properties:
+ *                       variants:
+ *                         type: array
+ *                         items:
+ *                           allOf:
+ *                             - $ref: "#/components/schemas/product_variant"
+ *                             - $ref: "#/components/schemas/product_variant_prices_fields"
  *             count:
  *               type: integer
  *               description: The total number of items available